move member LocationLPNList
